Added FormattingHelper.
Features:
- Completes (, [ and " with their counterparts.
- Indents the next line after { and completes it with }.
- Doesn’t indent after switch.
- Un-indents a "case" or "default" line to the same level as its switch
when : is typed.
- Indents the next line after a "case" or "default", but only if there
is no text after the :. Spaces are ignored.
- When hitting enter between two curly braces {}, the } is put two
lines down and the cursor on the line in between, indented.
- Typing a ), ] or " directly before another one when there isn’t one
missing just moves the cursor one step without adding another one.
- When there are already too many ), ] or }, they aren’t triggered by
their counterparts.
- I might have forgotten a smaller feature or two. It’s just so
complicated.

Summary: Tries to help the user with typing formatted code as well a
possible.